Remove matrix stack :(
Unfortunately the implementation I had for hierarchical
transformations in the scene graph wouldn't work in some
cases, (e.g. where I want a camera game object to
know its world co-ordinates). So because of this
I have gone back to computing the world matrix in 
the scene graph. This feels ugly.. But maybe its okay!